Version 0.1 of RPI TV Discord Bot
Added a setup, and createproduction command. Enables officers to create productions to post on a production channels, simultaneously creating a channel under a designated events category and allows volunteers to join via a button. Code is currently UNOPTIMIZED and only has the bare minimum. Also does not have any documentation on how to setup the bot.