Version 0.2 of RPI TV Discord Bot
Possibly ready for official use. Changed setup to be dropdown as opposed to just sending messages when we want to change categories/channels. Added editproduction command which enables officers to edit productions if they screwed up. Added archives option to setup so that when a production is past its due date, it moves into the archive category. This feature might be bugged due to limited testing so it's still up in the air.
Bugs may occur in any part of this code. Documentation on how to setup the bot is "finished." There might be some areas that can be worded better, but should be sufficient.

Added manual imports, fixed serviceaccountkey not being on gitignore, and changed 'interactionCreate' to Event.InteractionCreate, and any other applicable areas